+++ client.js
@@ -271,7 +271,7 @@ if (!command || command == "help" || command == "--help") {
 } else if (command == "download") {
 	console.log("Downloading mods...");
 	// get JSON data about releases
-	let res = syncRequest('GET', 'https://api.github.com/repos/Danielv123/factorioClusterioMod/releases', {"headers":{"User-Agent":"Fuck you for requiring user agents!"}});
+	let res = syncRequest('GET', 'https://api.github.com/repos/Danielv123/factorioClusterioMod/releases', {"headers":{"User-Agent":"factorioClusterio"}});
 	let url = JSON.parse(res.getBody())[0].assets[0].browser_download_url;
 	let name = JSON.parse(res.getBody())[0].assets[0].name;
 	if(url) {