+++ sharedPlugins/inventoryImports/isJson.js
@@ -2,7 +2,7 @@ module.exports = function(x){
 	if(x === false || x === true || x === null) return true; // === false is valid JSON, so is true and null
 	var isNotJson = false
 	try{
-		z = JSON.parse(x);
+		JSON.parse(x);
 	} catch (e) {
 		isNotJson = true;
 	}