+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Version 2.0.0
 
 - Fixed sslCert and sslPrivKey entries being ignored in the config.
 - Changed ssl key creation to be done at startup instead of on npm install.
+- Added error handling during master startup.
 
 Breaking Changes
 ----------------